Fix review suggestions
