+++ compiler-plugin/src/main/java/io/ballerina/lib/np/compilerplugin/PromptAsCodeCodeModificationTask.java
@@ -18,6 +18,10 @@
 
 package io.ballerina.lib.np.compilerplugin;
 
+import io.ballerina.compiler.api.SemanticModel;
+import io.ballerina.compiler.api.symbols.TypeReferenceTypeSymbol;
+import io.ballerina.compiler.api.symbols.TypeSymbol;
+import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
 import io.ballerina.compiler.syntax.tree.AbstractNodeFactory;
 import io.ballerina.compiler.syntax.tree.AnnotationNode;
 import io.ballerina.compiler.syntax.tree.DefaultableParameterNode;
@@ -34,6 +38,7 @@ import io.ballerina.compiler.syntax.tree.ImportPrefixNode;
 import io.ballerina.compiler.syntax.tree.MappingConstructorExpressionNode;
 import io.ballerina.compiler.syntax.tree.MappingFieldNode;
 import io.ballerina.compiler.syntax.tree.MetadataNode;
+import io.ballerina.compiler.syntax.tree.ModuleMemberDeclarationNode;
 import io.ballerina.compiler.syntax.tree.ModulePartNode;
 import io.ballerina.compiler.syntax.tree.NodeFactory;
 import io.ballerina.compiler.syntax.tree.NodeList;
@@ -41,13 +46,16 @@ import io.ballerina.compiler.syntax.tree.NodeParser;
 import io.ballerina.compiler.syntax.tree.ParameterNode;
 import io.ballerina.compiler.syntax.tree.QualifiedNameReferenceNode;
 import io.ballerina.compiler.syntax.tree.RequiredParameterNode;
+import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
 import io.ballerina.compiler.syntax.tree.SeparatedNodeList;
 import io.ballerina.compiler.syntax.tree.SimpleNameReferenceNode;
 import io.ballerina.compiler.syntax.tree.SpecificFieldNode;
 import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.compiler.syntax.tree.SyntaxTree;
 import io.ballerina.compiler.syntax.tree.Token;
 import io.ballerina.compiler.syntax.tree.TreeModifier;
 import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
+import io.ballerina.openapi.service.mapper.type.TypeMapper;
 import io.ballerina.projects.Document;
 import io.ballerina.projects.DocumentId;
 import io.ballerina.projects.Module;
@@ -56,7 +64,13 @@ import io.ballerina.projects.Package;
 import io.ballerina.projects.plugins.ModifierTask;
 import io.ballerina.projects.plugins.SourceModifierContext;
 import io.ballerina.tools.text.TextDocument;
+import io.swagger.v3.core.util.Json;
+import io.swagger.v3.core.util.OpenAPISchema2JsonSchema;
+import io.swagger.v3.oas.models.media.Schema;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createToken;
@@ -64,12 +78,16 @@ import static io.ballerina.compiler.syntax.tree.SyntaxKind.CLOSE_PAREN_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.DEFAULTABLE_PARAM;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.OPEN_PAREN_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.REQUIRED_PARAM;
+import static io.ballerina.lib.np.compilerplugin.Constants.BYTE;
 import static io.ballerina.lib.np.compilerplugin.Constants.CALL_LLM;
 import static io.ballerina.lib.np.compilerplugin.Constants.MODEL_VAR;
 import static io.ballerina.lib.np.compilerplugin.Constants.MODULE_NAME;
+import static io.ballerina.lib.np.compilerplugin.Constants.NUMBER;
 import static io.ballerina.lib.np.compilerplugin.Constants.ORG_NAME;
 import static io.ballerina.lib.np.compilerplugin.Constants.PROMPT_VAR;
 import static io.ballerina.lib.np.compilerplugin.Commons.hasLlmCallAnnotation;
+import static io.ballerina.lib.np.compilerplugin.Constants.STRING;
+import static io.ballerina.projects.util.ProjectConstants.EMPTY_STRING;
 
 /**
  * Code modification task to replace runtime prompt as code external functions with np:call.
@@ -85,6 +103,7 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
     private static final Token RIGHT_DOUBLE_ARROW = createToken(SyntaxKind.RIGHT_DOUBLE_ARROW_TOKEN);
     private static final Token COMMA = createToken(SyntaxKind.COMMA_TOKEN);
 
+    private static Optional<String> npPrefixIfImported = Optional.empty();
     private static final SimpleNameReferenceNode PROMPT_NAME_REF_NODE =
             NodeFactory.createSimpleNameReferenceNode(NodeFactory.createIdentifierToken(PROMPT_VAR));
     private static final SimpleNameReferenceNode MODEL_NAME_REF_NODE =
@@ -104,27 +123,58 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
             return;
         }
 
+        Map<String, String> typeSchemas = new HashMap<>();
         for (ModuleId moduleId : currentPackage.moduleIds()) {
             Module module = currentPackage.module(moduleId);
 
             for (DocumentId documentId: module.documentIds()) {
                 Document document = module.document(documentId);
+                processImportDeclarations(document);
+                processExternalFunctions(document, module, modifierContext, typeSchemas);
+            }
+
+            for (DocumentId documentId: module.documentIds()) {
+                Document document = module.document(documentId);
                 modifierContext.modifySourceFile(
-                        modifyDocument(document, analysisData), documentId);
+                        modifyDocument(document, typeSchemas), documentId);
             }
 
             for (DocumentId documentId: module.testDocumentIds()) {
                 Document document = module.document(documentId);
                 modifierContext.modifyTestSourceFile(
-                        modifyDocument(document, analysisData), documentId);
+                        modifyDocument(document, typeSchemas), documentId);
             }
         }
     }
 
-    private static TextDocument modifyDocument(Document document, CodeModifier.AnalysisData analysisData) {
+    private void processExternalFunctions(Document document, Module module,
+                                          SourceModifierContext modifierContext, Map<String, String> typeSchemas) {
+        if (npPrefixIfImported.isEmpty()) {
+            return;
+        }
+        SyntaxTree syntaxTree = document.syntaxTree();
+        ModulePartNode rootNode = syntaxTree.rootNode();
+        SemanticModel semanticModel = modifierContext.compilation().getSemanticModel(module.moduleId());
+        for (ModuleMemberDeclarationNode memberNode : rootNode.members()) {
+            if (!isExternalFunctionWithLlmCall(memberNode, npPrefixIfImported.get())) {
+                continue;
+            }
+
+            FunctionDefinitionNode functionDefinition = (FunctionDefinitionNode) memberNode;
+            extractAndStoreSchemas(semanticModel, functionDefinition, typeSchemas, this.analysisData.typeMapper);
+        }
+    }
+
+    private static void processImportDeclarations(Document document) {
+        ModulePartNode modulePartNode = document.syntaxTree().rootNode();
+        ImportDeclarationModifier importDeclarationModifier = new ImportDeclarationModifier();
+        modulePartNode.apply(importDeclarationModifier);
+    }
+
+    private static TextDocument modifyDocument(Document document, Map<String, String> typeSchemas) {
         ModulePartNode modulePartNode = document.syntaxTree().rootNode();
         FunctionModifier functionModifier = new FunctionModifier();
-        TypeDefinitionModifier typeDefinitionModifier = new TypeDefinitionModifier(analysisData);
+        TypeDefinitionModifier typeDefinitionModifier = new TypeDefinitionModifier(typeSchemas);
 
         ModulePartNode modifiedRoot = (ModulePartNode) modulePartNode.apply(functionModifier);
         modifiedRoot = modifiedRoot.modify(modifiedRoot.imports(), modifiedRoot.members(), modifiedRoot.eofToken());
@@ -135,9 +185,7 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
         return document.syntaxTree().modifyWith(finalRoot).textDocument();
     }
 
-    private static class FunctionModifier extends TreeModifier {
-
-        private Optional<String> npPrefixIfImported = Optional.empty();
+    private static class ImportDeclarationModifier extends TreeModifier {
 
         @Override
         public ImportDeclarationNode transform(ImportDeclarationNode importDeclarationNode) {
@@ -153,17 +201,20 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
             }
 
             Optional<ImportPrefixNode> prefix = importDeclarationNode.prefix();
-            this.npPrefixIfImported = Optional.of(prefix.isEmpty() ? MODULE_NAME : prefix.get().prefix().text());
+            npPrefixIfImported = Optional.of(prefix.isEmpty() ? MODULE_NAME : prefix.get().prefix().text());
             return importDeclarationNode;
         }
+    }
+
+    private static class FunctionModifier extends TreeModifier {
 
         @Override
         public FunctionDefinitionNode transform(FunctionDefinitionNode functionDefinition) {
-            if (this.npPrefixIfImported.isEmpty()) {
+            if (npPrefixIfImported.isEmpty()) {
                 return functionDefinition;
             }
 
-            String npPrefix = this.npPrefixIfImported.get();
+            String npPrefix = npPrefixIfImported.get();
 
             FunctionBodyNode functionBodyNode = functionDefinition.functionBody();
 
@@ -230,22 +281,25 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
 
     private static class TypeDefinitionModifier extends TreeModifier {
 
-        private final CodeModifier.AnalysisData analysisData;
+        private final Map<String, String> typeSchemas;
 
-        TypeDefinitionModifier(CodeModifier.AnalysisData analysisData) {
-            this.analysisData = analysisData;
+        TypeDefinitionModifier(Map<String, String> typeSchemas) {
+            this.typeSchemas = typeSchemas;
         }
 
         @Override
         public TypeDefinitionNode transform(TypeDefinitionNode typeDefinitionNode) {
+            if (npPrefixIfImported.isEmpty()) {
+                return typeDefinitionNode;
+            }
             String typeName = typeDefinitionNode.typeName().text();
 
-            if (!analysisData.typeSchemas.containsKey(typeName)) {
+            if (!this.typeSchemas.containsKey(typeName)) {
                 return typeDefinitionNode;
             }
 
             MetadataNode updatedMetadataNode =
-                                            updateMetadata(typeDefinitionNode, analysisData.typeSchemas.get(typeName));
+                                        updateMetadata(typeDefinitionNode, typeSchemas.get(typeName));
             return typeDefinitionNode.modify().withMetadata(updatedMetadataNode).apply();
         }
 
@@ -303,5 +357,140 @@ public class PromptAsCodeCodeModificationTask implements ModifierTask<SourceModi
                 AbstractNodeFactory.createToken(SyntaxKind.COLON_TOKEN),
                 NodeParser.parseExpression(jsonSchema)
         );
+    }
+
+    private boolean isExternalFunctionWithLlmCall(ModuleMemberDeclarationNode memberNode, String npModulePrefixStr) {
+        if (!(memberNode instanceof FunctionDefinitionNode functionDefinition)) {
+            return false;
+        }
+        return functionDefinition.functionBody() instanceof ExternalFunctionBodyNode externalFunctionBodyNode
+                && hasLlmCallAnnotation(externalFunctionBodyNode, npModulePrefixStr);
+    }
+
+    private void extractAndStoreSchemas(SemanticModel semanticModel, FunctionDefinitionNode functionDefinition,
+                                        Map<String, String> typeSchemas, TypeMapper typeMapper) {
+        Optional<ReturnTypeDescriptorNode> returnTypeNodeOpt = functionDefinition.functionSignature().returnTypeDesc();
+        if (returnTypeNodeOpt.isEmpty()) {
+            return;
+        }
+
+        ReturnTypeDescriptorNode returnTypeNode = returnTypeNodeOpt.get();
+        Optional<TypeSymbol> typeSymbolOpt = semanticModel.type(returnTypeNode.type().lineRange());
+        if (typeSymbolOpt.isEmpty()) {
+            return;
+        }
+
+        TypeSymbol typeSymbol = typeSymbolOpt.get();
+        if (typeSymbol instanceof UnionTypeSymbol unionTypeSymbol) {
+            for (TypeSymbol memberType : unionTypeSymbol.memberTypeDescriptors()) {
+                if (memberType instanceof TypeReferenceTypeSymbol typeReferenceTypeSymbol) {
+                    Schema schema = typeMapper.getSchema(typeReferenceTypeSymbol);
+                    typeSchemas.put(typeReferenceTypeSymbol.definition().getName().get(), getJsonSchema(schema));
+                }
+            }
+        }
+    }
+
+    @SuppressWarnings("rawtypes")
+    private static String getJsonSchema(Schema schema) {
+        modifySchema(schema);
+        OpenAPISchema2JsonSchema openAPISchema2JsonSchema = new OpenAPISchema2JsonSchema();
+        openAPISchema2JsonSchema.process(schema);
+        String newLineRegex = "\\R";
+        String jsonCompressionRegex = "\\s*([{}\\[\\]:,])\\s*";
+        return Json.pretty(schema.getJsonSchema())
+                .replaceAll(newLineRegex, EMPTY_STRING)
+                .replaceAll(jsonCompressionRegex, "$1");
+    }
+
+    @SuppressWarnings({"unchecked", "rawtypes"})
+    private static void modifySchema(Schema schema) {
+        if (schema == null) {
+            return;
+        }
+        modifySchema(schema.getItems());
+        modifySchema(schema.getNot());
+
+        Map<String, Schema> properties = schema.getProperties();
+        if (properties != null) {
+            properties.values().forEach(PromptAsCodeCodeModificationTask::modifySchema);
+        }
+
+        List<Schema> allOf = schema.getAllOf();
+        if (allOf != null) {
+            schema.setType(null);
+            allOf.forEach(PromptAsCodeCodeModificationTask::modifySchema);
+        }
+
+        List<Schema> anyOf = schema.getAnyOf();
+        if (anyOf != null) {
+            schema.setType(null);
+            anyOf.forEach(PromptAsCodeCodeModificationTask::modifySchema);
+        }
+
+        List<Schema> oneOf = schema.getOneOf();
+        if (oneOf != null) {
+            schema.setType(null);
+            oneOf.forEach(PromptAsCodeCodeModificationTask::modifySchema);
+        }
+
+        // Override default ballerina byte to json schema mapping
+        if (BYTE.equals(schema.getFormat()) && STRING.equals(schema.getType())) {
+            schema.setFormat(null);
+            schema.setType(NUMBER);
+        }
+        removeUnwantedFields(schema);
+    }
+
+    @SuppressWarnings({"rawtypes", "unchecked"})
+    private static void removeUnwantedFields(Schema schema) {
+        schema.setSpecVersion(null);
+        schema.setSpecVersion(null);
+        schema.setContains(null);
+        schema.set$id(null);
+        schema.set$schema(null);
+        schema.set$anchor(null);
+        schema.setExclusiveMaximumValue(null);
+        schema.setExclusiveMinimumValue(null);
+        schema.setDiscriminator(null);
+        schema.setTitle(null);
+        schema.setMaximum(null);
+        schema.setExclusiveMaximum(null);
+        schema.setMinimum(null);
+        schema.setExclusiveMinimum(null);
+        schema.setMaxLength(null);
+        schema.setMinLength(null);
+        schema.setMaxItems(null);
+        schema.setMinItems(null);
+        schema.setMaxProperties(null);
+        schema.setMinProperties(null);
+        schema.setAdditionalProperties(null);
+        schema.setAdditionalProperties(null);
+        schema.set$ref(null);
+        schema.set$ref(null);
+        schema.setReadOnly(null);
+        schema.setWriteOnly(null);
+        schema.setExample(null);
+        schema.setExample(null);
+        schema.setExternalDocs(null);
+        schema.setDeprecated(null);
+        schema.setPrefixItems(null);
+        schema.setContentEncoding(null);
+        schema.setContentMediaType(null);
+        schema.setContentSchema(null);
+        schema.setPropertyNames(null);
+        schema.setUnevaluatedProperties(null);
+        schema.setMaxContains(null);
+        schema.setMinContains(null);
+        schema.setAdditionalItems(null);
+        schema.setUnevaluatedItems(null);
+        schema.setIf(null);
+        schema.setElse(null);
+        schema.setThen(null);
+        schema.setDependentSchemas(null);
+        schema.set$comment(null);
+        schema.setExamples(null);
+        schema.setExtensions(null);
+        schema.setConst(null);
     }
 }
